Move update check to app startup, remove /upgrade TUI command
Instead of a /upgrade slash command, the TUI now checks for updates
automatically when the app opens. If a newer version is available, a
warning toast appears with the exact command to run (e.g.
'npm install -g @pensar/apex@latest'). The check is async and
non-blocking so it doesn't delay startup.

- Remove /upgrade command and toast plumbing from command-registry
- Remove toast from AppCommandContext and CommandProvider
- Add update check useEffect in AppContent (inside ToastProvider)
